Adding version 1.4.1
git-svn-id: https://plugins.svn.wordpress.org/custom-field-images/trunk@62407 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/admin.php
@@ -19,9 +19,13 @@ class cfImgAdmin extends cfImg {
 	function activate() {
 		$ver = 	get_option('cfi_version');
 
-		if ($ver === '1.3')
+		if ($ver === '1.4')
+			return;
+
+		if ($ver === '1.3') {
 			$show_in = get_option('cfi_options');
-		else {
+			update_option('cfi_version', '1.4');
+		} else {
 			$show_in = get_option('cfi-show-in');
 			delete_option('cfi-show-in');
 		}
@@ -31,8 +35,6 @@ class cfImgAdmin extends cfImg {
 
 		   add_option('cfi_options', $this->options) or
 		update_option('cfi_options', $this->options);
-
-		update_option('cfi_version', '1.4');
 	}
 
 	function upgrade() {
